+++ tests/hash_test.py
@@ -1,9 +1,9 @@
-import os
+from photoshell.hash import hash_file
 
-from photoshell import util
+import os
 
 
 def test_hash_file(tmpdir):
     tmpdir.join('file.test').write("Test")
-    assert (util.hash_file(os.path.join(tmpdir.strpath, 'file.test')) ==
+    assert (hash_file(os.path.join(tmpdir.strpath, 'file.test')) ==
             '640ab2bae07bedc4c163f679a746f7ab7fb5d1fa')
